Get rid of a warning.

+++ server/src/TPlayerProps.cpp
@@ -313,7 +313,7 @@ void TPlayer::setProps(CString& pPacket, bool pForward, bool pForwardToSelf, TPl
 				maxPower = pPacket.readGUChar();
 				maxPower = clip(maxPower, 0, settings->getInt("heartlimit", 20));
 				maxPower = clip(maxPower, 0, 20);
-				power = maxPower;
+				power = (float)maxPower;
 				levelBuff >> (char)PLPROP_CURPOWER << getProp(PLPROP_CURPOWER);
 				selfBuff >> (char)PLPROP_CURPOWER << getProp(PLPROP_CURPOWER);
 			break;
